Wikidata.Summarizing.Langgraph/requirements.txt
langgraph==0.4.3
langchain==0.3.25
langchain-openai==0.3.16
langchain-community==0.3.23
pydantic==2.11.4
typing-extensions==4.13.2
python-dotenv==1.1.0
requests==2.32.3
azure-identity==1.15.0
